fixup indexing include orders
