+++ src/test/unit/model/indexing/assign_test.cpp
@@ -1,10 +1,8 @@
+#include <stan/model/indexing.hpp>
+#include <gtest/gtest.h>
 #include <iostream>
 #include <stdexcept>
 #include <vector>
-#include <stan/model/indexing/assign.hpp>
-#include <stan/model/indexing/rvalue.hpp>
-#include <stan/math/rev.hpp>
-#include <gtest/gtest.h>
 
 using Eigen::Dynamic;
 using Eigen::Matrix;
