+++ src/test/unit/model/indexing/assign_varmat_test.cpp
@@ -1,13 +1,10 @@
-#include <iostream>
-#include <stdexcept>
-#include <vector>
-#include <stan/model/indexing/assign_varmat.hpp>
-#include <stan/model/indexing/assign.hpp>
-#include <stan/model/indexing/rvalue.hpp>
-#include <stan/math/rev.hpp>
+#include <stan/model/indexing.hpp>
 #include <test/unit/util.hpp>
 #include <test/unit/model/indexing/util.hpp>
 #include <gtest/gtest.h>
+#include <iostream>
+#include <stdexcept>
+#include <vector>
 
 using Eigen::Dynamic;
 using Eigen::Matrix;
